requirements.txt
Django==2.1
django-cors-headers==2.5.0
django-filter==2.1.0
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
